chore: move metrics api into base Unleash
