+++ src/test/impact-metrics/metric-client.test.ts
@@ -1,4 +1,4 @@
-import { MetricsAPI } from '../../impact-metrics/metric-client';
+import { MetricsAPI } from '../../impact-metrics/metric-api';
 
 import test from 'ava';
 import Client from '../../client';
